Battery indicator

+++ Assets/Scripts/CommandPattern/CommandHandler.cs
@@ -1,28 +1,39 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.UI;
 
 public class CommandHandler
 {
     public List<ICommand> commandList = new List<ICommand>();
     public List<Vector3Int> selectedLocations = new List<Vector3Int>();
     public List<float> riskValues = new List<float>();
+    public int batteryLevel = 12;
     public float accumulatedRisk = 0;
     public int index;
+
     // For the risk graph
     private Window_Graph windowGraph;
 
     public void AddCommand(ICommand command)
-    {       
-        {if (index < commandList.Count)
-            commandList.RemoveRange(index, commandList.Count - index);
-            selectedLocations.RemoveRange(index, selectedLocations.Count - index);
+    {
+        if (batteryLevel > 0)
+        {
+            {if (index < commandList.Count)
+                commandList.RemoveRange(index, commandList.Count - index);
+                selectedLocations.RemoveRange(index, selectedLocations.Count - index);
 
-        commandList.Add(command);
-        selectedLocations.Add(command.clickedLocation);
-        addRisk(command.clickedLocation);
-        command.Execute();
-        index++;
+            commandList.Add(command);
+            selectedLocations.Add(command.clickedLocation);
+            addRisk(command.clickedLocation);
+            removeEnergy(1);
+            command.Execute();
+            index++;
+            }
+        }
+        else
+        {
+            Debug.Log("Battery is empty");
         }
     }
 
@@ -35,6 +46,7 @@ public class CommandHandler
             commandList[index - 1].Undo();
             selectedLocations.RemoveAt(index - 1);
             removeRisk(commandList[index - 1].clickedLocation); // not sure if index is 1 or 0
+            addEnergy(1);
             index--;
         }
         Debug.Log("Command removed");
@@ -60,5 +72,48 @@ public class CommandHandler
         windowGraph.ShowGraph(riskValues);
     }
     
+
+    void addEnergy(int energy)
+    {
+        batteryLevel += energy;
+        Debug.Log($"adding energy, current level: {batteryLevel}");
+        if(batteryLevel < 7)
+        {
+            Image batteryLevelText = GameObject.Find("Full").GetComponent<Image>();
+            batteryLevelText.sprite = Resources.Load<Sprite>("Batteries/Half");
+            if (batteryLevel < 3)
+            {
+                batteryLevelText.sprite = Resources.Load<Sprite>("Batteries/Low");
+                if (batteryLevel <= 0)
+                {
+                    batteryLevelText.sprite = Resources.Load<Sprite>("Batteries/Zero");
+                }
+            }
+        }
+
+    }
+    void removeEnergy(int energy)
+    {
+        batteryLevel -= energy;
+        if(batteryLevel < 7)
+        {
+            Image batteryLevelText = GameObject.Find("Full").GetComponent<Image>();
+            batteryLevelText.sprite = Resources.Load<Sprite>("Batteries/Half");
+            if (batteryLevel < 3)
+            {
+                batteryLevelText.sprite = Resources.Load<Sprite>("Batteries/Low");
+                if (batteryLevel <= 0)
+                {
+                    batteryLevelText.sprite = Resources.Load<Sprite>("Batteries/Zero");
+                }
+            }
+
+        }
+        Debug.Log($"removing energy, current level: {batteryLevel}");
+    }
+    void showBatteryIcon()
+    {
+        // TODO
+    }
 }
